Move lipyphilic.lib to lipyphilic.analysis
