+++ docs/reference/lib/registration.rst
@@ -1,1 +1,1 @@
-.. automodule:: lipyphilic.lib.registration
+.. automodule:: lipyphilic.analysis.registration
