+++ docs/reference/lib/z_pos.rst
@@ -1,1 +1,1 @@
-.. automodule:: lipyphilic.lib.z_positions
+.. automodule:: lipyphilic.analysis.z_positions
