+++ tests/lipyphilic/lib/test_area_per_lipid.py
@@ -10,7 +10,7 @@ from lipyphilic._simple_systems.simple_systems import (
     HEX_LAT_OVERLAP,
     TRICLINIC,
 )
-from lipyphilic.lib.area_per_lipid import AreaPerLipid
+from lipyphilic.analysis.area_per_lipid import AreaPerLipid
 from lipyphilic.plotting import ProjectionPlot
 
 mpl.use("Agg")
