+++ tests/lipyphilic/lib/test_order_parameter.py
@@ -5,7 +5,7 @@ from numpy.testing._private.utils import assert_allclose, assert_array_almost_eq
 import pytest
 
 from lipyphilic._simple_systems.simple_systems import HEX_LAT, ONE_CHOL, ONE_CHOL_TRAJ
-from lipyphilic.lib.order_parameter import SCC
+from lipyphilic.analysis.order_parameter import SCC
 from lipyphilic.plotting import ProjectionPlot
 
 mpl.use("Agg")
